Modified business rules interface to have a little more granularity: beforeSave(), beforeSaveCommit(), afterSaveCommit(). Fixed a few lingering problems related to the switch from Long to Integer for ID types.
git-svn-id: https://svn.code.sf.net/p/specify/code/trunk/Specify@2608 ce9bf4bf-376f-1848-a496-a1c9fe3b1a39

+++ src/edu/ku/brc/specify/config/Scriptlet.java
@@ -431,7 +431,7 @@ public class Scriptlet extends JRDefaultScriptlet
      * @param colEvId
      * @return
      */
-    public String getCollectors(final Long colEvId)
+    public String getCollectors(final Integer colEvId)
     {
         DataProviderSessionIFace session = DataProviderFactory.getInstance().createSession();
         //System.out.println(colEvId);
